+++ src/main/java/handler/ResponseHandler.java
@@ -1,4 +1,4 @@
-package webserver;
+package handler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
